Split queue_entry.py into smaller pieces
  - Each QueueEntry object is in its own file named <queue_entry_name>queue_entry
    snake cased name with a queue_entry suffix

  - The QueueEntry object is named <QueueEntryName>QueueEntry
    camel cased name with a QueueEntry suffix (or simply the camel cased filename)


  - Created queue_entry sub-module that automatically imports any *_queue_entry.py
    file in addition to the constants and base classed defined in base_queue_entry

  - Import of queue_entry submodule is more or less a drop in replacement of the
    previous queue_etnry.py

  - New queue entrie can simply be added by adding a new <name>_queue_entry.py file
