+++ mxcubecore/HardwareObjects/QueueManager.py
@@ -9,9 +9,10 @@ documentation for the queue_entry module for more information.
 """
 import logging
 import gevent
-from mxcubecore.HardwareObjects import base_queue_entry, queue_entry
+from mxcubecore.HardwareObjects import queue_entry
 from mxcubecore.BaseHardwareObjects import HardwareObject
-from mxcubecore.HardwareObjects.base_queue_entry import QUEUE_ENTRY_STATUS
+from mxcubecore.HardwareObjects.queue_entry.base_queue_entry import QUEUE_ENTRY_STATUS
+from mxcubecore.HardwareObjects.queue_entry import base_queue_entry
 
 QueueEntryContainer = base_queue_entry.QueueEntryContainer
